+++ src/store/store.ts
@@ -12,3 +12,4 @@ export const store = () => {
 
 export type RootState = ReturnType<typeof rootReducer>
 export type AppStore = ReturnType<typeof store>
+export type AppDispatch = AppStore['dispatch']
